refactor: organizing code and using atomFamily
