+++ apps/namadillo/src/atoms/integrations/index.ts
@@ -1,1 +1,3 @@
 export * from "./atoms";
+export * from "./functions";
+export * from "./services";
