--- apps/namadillo/src/atoms/registry/index.ts
@@ -1,2 +1,0 @@
-export * from "./functions";
-export * from "./services";
